Value: fix+test crash when default value is bad and code fix returns undef...
this fix a bug seen in the patched version if lcdproc model done for Debian.
The model is something like:

#   compute
#     formula="my $l = '/dev/lcd-imon'; -e $l ? $l : '/dev/lcd0';"
#     use_eval=1
#     allow_override=1 -
#   warn_if:not_lcd_imon
#     code="my $l = '/dev/lcd-imon';defined $_ and -e $l and $_ ne $l ;"
#     msg="imon device does not use /dev/lcd-imon link."
#     fix="$_ = undef;"
#   warn_unless:found_device_file
#     code="defined $_ ? -e : 1"
#     msg="missing imon device file"
#     fix="$_ = undef;"

When lcdproc kernel module is not loaded, both /dev/lcd0 and /dev/lcd-imon
are missing so code fix returns undef . the code tried to notify change
from defaut value to default value which is fatal to notify_change. This patch
prevents this error