Model: allow inecjt snippet in other class
Until now a class snippet could only augment the class it was stored
with. For instance:

  Foo.pl # Foo model
  Foo.d/stuff.pl # snippet

Until now, the data stored in stuff.pl could only augment Foo or a class
alreay loaded.

Now the snippet are stored so the snippet can augment a class which is
not yet loaded.

This is required for Dpkg model. Where we have:

  Dpkg.pl
  Dpkg.d/source-name.pl
  Dpkg/Copyright.pl

Here Dpkg and Dpkg::Copyright can be both model root (respectively with
'cme edit dpkg' and 'cme edit dpkg-copyright'

Dpkg model, requires Dpkg::Copyright. In this case, we want data in
Dpkg::Copyright to be computed from Data in Dpkg.

When Dpkg::Copyright is root, this data is non-existent, so compute
cannot be setup.

Solution: add a model snipped associated with Dpkg (in Dpkg.d). This
snippet is load with Dpkg, but can be applied to Dpkg::Copyright only
when this class is loaded (when copyright node is opened by user).

Storing snippet in model enables this solution.