Revert "rework compute parser interface"
This reverts commit 21e019953e5323770e2d60e3997d209fdb498c02.

This commit broke tests of libconfig-model-dpkg-perl. The trick is the
need_quote parameters. It's needed during use_eval depending on the type
of variable read. It's tricky to get right and the re-org of this commit
did break it. This re-org is no longer necessary and not worth my time.