Node: all 'all' option to get_element_names
Mostly useful when used in a backend.

why is option 'all' needed: This fixes an issue seen with Dpkg
backend. get_element_name triggers warp, which means that test-control's
warp is triggered which trigger a read on its warp master. In other
word, control file content is read before another warp master (compat)
is read, so the default value of compat is returned at this stage
instead of the value from compat file. Hence dpkg/conrol model issues a
warning about a value mistach with compat when there's none. So this
loop should list elements from model, read the data but check if the
element is available or not.