notify of hash addition when needed
There's no need to notify user of a new hash key only of writing empty
value is permitted.

When writing an empty value is not permitted, user will be informed of
new data in the hash for a specific key. Hence user will get
notification of new key and new value in ony one message.

Tests needed to be fixed because less notifications are sent.